Get a call stack, because deferred panics suck
A side-effect of the options pattern is that we might not be able to
detect an error until the time when the options are applied. These
moments may be a bit further apart, making it harder to debug.

There's only one place I foresee this happening, and that's with the
exactly-once requirement for WithRun/WithStartStop (providing both is an
error, as is calling either one twice).

As a way to head this off, I'll introduce this little ditty to capture a
fragment fo the stack and hopefully make it just a touch easier.