Move most internal errors into their own namespace
Upon consideration, I think I'd like the e2e tests to be able to use
these, as one of the only allowances for e2e to use any internal
packages.

That said, I don't think we should ever be checking for
errPrematureChannelClose, so I've kept it as an private component detail.