ImplCheckReady policy change: Returned errors are bad
Previously, the policy was that if the user-generated code returned an
error, then we'd log the error but we would not consider that error as
bad enough to fail the startup.

This made for some rather messy policy matters, including a mismatch in
the default behaviors of what an error means to use.

The new policy is that if CheckReady returns an error, then that means
the component will not successfully come online, and we should abandon
the startup attempt.

We may add in a policy setting later on to control this. I'm still
contemplating what that'd look like, and don't want to inflate this
commit any further.