go1.25 upgrade; synctest promotion breaks things
synctest is no longer experimental, but it's signature changed.

It also had a breaking behavioral change: if the main goroutine for the
synctest bubble exits, and any coroutines inside the bubble are still
durably blocked, it now panics. (Previously, it let those finish running
on their own.)

This is certainly a positive change, but I didn't write my earlier code
under that assumption, so I threw in a hacky workaround for this change.